refactor(core): 迁移请求日志上下文到 shared/observability
